refactor(scanner): consolidate future boxing and address review nits
Keep the async-layout recursion boxing at the call sites (explain_plan, the
point-lookup scan builder, remap_indices) rather than inside create_plan /
remap_index: boxing *inside* those methods turns the future's `Send` check
into a `Box<Future>: Send` trait obligation that overflows the solver through
the moka cache types (E0275 in downstream crates). Dropped the now-redundant
`plan_lookup` box in the bench — the scan-builder boxes already keep that
chain shallow. Reworded the boxing comments to explain the mechanism.

Review nits:
- `EvaluatedIndex::without_rows` takes `&RowAddrTreeMap` (the block-list
  invariant moves to the caller) instead of asserting internally.
- Move `translate_addr_treemap_to_row_ids` to `dataset::rowids`, next to
  `load_row_id_sequence`, instead of widening its visibility in scalar_index.
