+++ rust/lance/benches/mem_wal/point_lookup/mem_wal_point_lookup_bench.rs
@@ -431,9 +431,9 @@ async fn run_lookup(args: &Args) -> Result<serde_json::Value> {
 
         for (qi, &id) in ids.iter().enumerate() {
             let t0 = Instant::now();
-            // Box this deep future: overlay-aware index masking pushes its layout past rustc's
-            // default recursion limit when inlined here.
-            let plan = Box::pin(planner.plan_lookup(&[ScalarValue::Int64(Some(id))], None)).await?;
+            let plan = planner
+                .plan_lookup(&[ScalarValue::Int64(Some(id))], None)
+                .await?;
             let stream = plan.execute(0, task_ctx.clone())?;
             let batches: Vec<RecordBatch> = stream.try_collect().await?;
             let elapsed_us = t0.elapsed().as_micros() as f64;
